Don't enabled webpack watch for prod, fixes inability to run yarn deploy
Updated the webpack config to take an argument whether or not to use watch, and assume true if no argument is supplied.

This makes it so that watch is not enabled on production runs, which is used when doing yarn deploy. However, with watch enabled it would not return from watch and create a commit of the build output like it's supposed to.

So watch is now only enabled for yarn build.