Webpack now uses sourcemapping to show original JSX/JS source code in debugger
Previously when generating a build of our website, the chrome/firefox debugger would show the compilation generated .js files that are not meant to be human read.

Instead I have enabled source mapping so that the debugger shows the original code.

This makes finding error locations easier, and makes debugging significantly easier.