Added webpack-dev-server to build to localhost and use auto-build & auto-refresh
Webpack dev server allows us to run yarn start to build the website and deploy it to localhost.

It will automatically detect file changes and rebuild, and the command has been supplied with the --hot command so that it not only auto-refreshes when files are changed, but it only replaces elements that have been modified instead of doing a full refresh.