Added functions to simulate entering info to sign up form
Since we don't have time to implement a database and handle network requests/email handing/authentication of registration and login, we just want to demonstrate the idea.

So I have added a function that displays typing in credentials into the signup form as an illustration.

This will be expanded as we simulate log in by changing the login/sing up buttons, and show a way to get to the stream page.