Don't logout when clicking sign up button again
The page would reload and reset to the original url upon clicking the submit button a second time.

I have changed this input to a standard button so that it does not attempt to submit, and rather just follows the operations of the onclick function used, which sets the query only if there is not already one.

If the user is already logged in it will still refresh, but the url will not change.