Update login/logout to use react state/props
We are not doing actual logging in and registration due to time restraints. We simply want to demo the difference between logging in and logged out to show that the user needs to sign up to access the stream.

Originally we were using a query string, but instead of managing and editing the url it is better to just make use of React. The shared app component stores the logged in variable and passes down that value as a prop to the components that need to check if the user is "logged in" (such as the header componet), and passes a function down to change the status of the user logged in state for buttons that modify this.

This makes things more immersive than having an obvious visible query that handles the status.